# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LifeBridge AI",
    "version": "1.1.0"
  },
  "paths": {
...
    "/api/v1/feed": {
      "get": {
        "tags": [
          "feed"
        ],
        "summary": "Get Feed",
        "operationId": "get_feed_api_v1_feed_get",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "location",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "Location"
            }
          },
          {
            "name": "verified_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Verified Only"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 5000,
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FeedItemOut"
                  },
                  "title": "Response Get Feed Api V1 Feed Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/feed/{item_id}": {
      "get": {
        "tags": [
          "feed"
        ],
        "summary": "Get Feed Item",
        "operationId": "get_feed_item_api_v1_feed__item_id__get",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedItemOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recommendations": {
      "post": {
        "tags": [
          "feed"
        ],
        "summary": "Recommendations",
        "operationId": "recommendations_api_v1_recommendations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecommendationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RecommendationOut"
                  },
                  "type": "array",
                  "title": "Response Recommendations Api V1 Recommendations Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/profile": {
      "get": {
        "tags": [
          "profile"
        ],
        "summary": "Read Profile",
        "operationId": "read_profile_api_v1_profile_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "profile"
        ],
        "summary": "Delete Account",
        "operationId": "delete_account_api_v1_profile_delete",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "profile"
        ],
        "summary": "Update Profile",
        "operationId": "update_profile_api_v1_profile_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserProfileUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/profile/export": {
      "get": {
        "tags": [
          "profile"
        ],
        "summary": "Export Profile",
        "operationId": "export_profile_api_v1_profile_export_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/saved": {
      "get": {
        "tags": [
          "saved"
        ],
        "summary": "List Saved",
        "operationId": "list_saved_api_v1_saved_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FeedItemOut"
                  },
                  "type": "array",
                  "title": "Response List Saved Api V1 Saved Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "saved"
        ],
        "summary": "Save Item",
        "operationId": "save_item_api_v1_saved_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveItemRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/saved/{feed_item_id}": {
      "delete": {
        "tags": [
          "saved"
        ],
        "summary": "Remove Saved",
        "operationId": "remove_saved_api_v1_saved__feed_item_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "feed_item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Feed Item Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FeedItemOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "external_id": {
            "type": "string",
            "title": "External Id"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "source_name": {
            "type": "string",
            "title": "Source Name"
          },
          "source_url": {
            "type": "string",
            "title": "Source Url"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "published_at": {
            "type": "string",
            "format": "date-time",
            "title": "Published At"
          },
          "collected_at": {
            "type": "string",
            "format": "date-time",
            "title": "Collected At"
          },
          "last_checked_at": {
            "type": "string",
            "format": "date-time",
            "title": "Last Checked At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "country_code": {
            "type": "string",
            "title": "Country Code",
            "default": ""
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "tags": {
            "type": "string",
            "title": "Tags"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "verification_status": {
            "type": "string",
            "title": "Verification Status"
          },
          "source_reliability": {
            "type": "number",
            "title": "Source Reliability"
          },
          "funding_type": {
            "type": "string",
            "title": "Funding Type",
            "default": ""
          },
          "study_level": {
            "type": "string",
            "title": "Study Level",
            "default": ""
          },
          "employment_type": {
            "type": "string",
            "title": "Employment Type",
            "default": ""
          },
          "salary_text": {
            "type": "string",
            "title": "Salary Text",
            "default": ""
          },
          "eligibility": {
            "type": "string",
            "title": "Eligibility",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "id",
          "external_id",
          "category",
          "title",
          "summary",
          "source_name",
          "source_url",
          "published_at",
          "collected_at",
          "last_checked_at",
          "updated_at",
          "expires_at",
          "location",
          "tags",
          "severity",
          "verification_status",
          "source_reliability"
        ],
        "title": "FeedItemOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RecommendationOut": {
        "properties": {
          "item": {
            "$ref": "#/components/schemas/FeedItemOut"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reasons"
          }
        },
        "type": "object",
        "required": [
          "item",
          "score",
          "reasons"
        ],
        "title": "RecommendationOut"
      },
      "RecommendationRequest": {
        "properties": {
          "skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 50,
            "title": "Skills"
          },
          "preferred_categories": {
            "items": {
              "type": "string",
              "enum": [
                "job",
                "scholarship",
                "disaster",
                "weather",
                "service",
                "safety",
                "learning"
              ]
            },
            "type": "array",
            "maxItems": 10,
            "title": "Preferred Categories"
          },
          "country": {
            "type": "string",
            "maxLength": 100,
            "title": "Country",
            "default": ""
          },
          "city": {
            "type": "string",
            "maxLength": 100,
            "title": "City",
            "default": ""
          },
          "study_level": {
            "type": "string",
            "maxLength": 80,
            "title": "Study Level",
            "default": ""
          },
          "field_of_study": {
            "type": "string",
            "maxLength": 120,
            "title": "Field Of Study",
            "default": ""
          },
          "interests": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 30,
            "title": "Interests"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 90.0,
                "minimum": -90.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 180.0,
                "minimum": -180.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "limit": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Limit",
            "default": 20
          }
        },
        "type": "object",
        "title": "RecommendationRequest"
      },
...
      "SaveItemRequest": {
        "properties": {
          "feed_item_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Feed Item Id"
          },
          "reminder_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reminder At"
          }
        },
        "type": "object",
        "required": [
          "feed_item_id"
        ],
        "title": "SaveItemRequest"
      },
...
      "UserProfileOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "country": {
            "type": "string",
            "title": "Country"
          },
          "city": {
            "type": "string",
            "title": "City"
          },
          "study_level": {
            "type": "string",
            "title": "Study Level"
          },
          "field_of_study": {
            "type": "string",
            "title": "Field Of Study"
          },
          "skills": {
            "type": "string",
            "title": "Skills"
          },
          "interests": {
            "type": "string",
            "title": "Interests"
          },
          "preferred_categories": {
            "type": "string",
            "title": "Preferred Categories"
          },
          "accessibility_preferences": {
            "type": "string",
            "title": "Accessibility Preferences"
          },
          "language": {
            "type": "string",
            "title": "Language"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "display_name",
          "country",
          "city",
          "study_level",
          "field_of_study",
          "skills",
          "interests",
          "preferred_categories",
          "accessibility_preferences",
          "language"
        ],
        "title": "UserProfileOut"
      },
      "UserProfileUpdate": {
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "country": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Country"
          },
          "city": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "study_level": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "Study Level"
          },
          "field_of_study": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Field Of Study"
          },
          "skills": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array",
                "maxItems": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "Skills"
          },
          "interests": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array",
                "maxItems": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Interests"
          },
          "preferred_categories": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "enum": [
                    "job",
                    "scholarship",
                    "disaster",
                    "weather",
                    "service",
                    "safety",
                    "learning"
                  ]
                },
                "type": "array",
                "maxItems": 10
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Categories"
          },
          "accessibility_preferences": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array",
                "maxItems": 30
              },
              {
                "type": "null"
              }
            ],
            "title": "Accessibility Preferences"
          },
          "language": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          }
        },
        "type": "object",
        "title": "UserProfileUpdate"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}